feat:开源版部署时支持一键初始化插件 #7556
